Fix saves dialog condition
Issue was that I'd manually editted the "show only if" condition in the AIP and it
wasn't correctly working with the "Back" button. This is now fixed.

Second issue is that it seems you can't have multiple sequential "show only if"
dialogs at the same point in the sequence - if you do, only the first one triggered
will show.

As a workaround, the Tobii/Mouse info dialog now comes one step later.