Hide "ignore warning" checkbox
Checkbox only appears if you click warning image.
This lets us give people a backup option if necessary but hopefully isn't
discoverable enough for people to shoot themselves in the foot easily.

Have also removed the forge "ignore warning" checkbox since it's
likely the rest of the installation will fail in that case - need to check