Fix installer for non-admin user
Installer was looking in administrator's APPDATA directory (authenticated user) instead of
the logged-in user who had launched the installer. The issue occurred because the deferred
Custom Actions were run as the authenticated admin user. Earlier on in the installation process
we can query the AI property "AppDataFolder" and then pass it into the Custom Actions.

This change opens up the opportunity to give user the chance to change the .minecraft
installation directory rather than assuming the default (but the installer would need to
have extra UI screens to enable this)

Next time we will avoid any installation straight into APPDATA, and perhaps install things
at first runtime?

Various changes:
- Move helper methods to separate classes
- Cache appdata info from / into installer property
- Pass this path via CustomActionData
- Extract path logic (different folders) to single struct
- Manual test code