Aggressively check/force focus for Ghost before automating
There was an issue on one deployment that we think was caused by some
pre-loaded bloatware that was grabbing focus. The Ghost automation
only worked if you manually clicked to re-focus the app, because the
AHK script had hung on "WinWaitActive". This version separates the
WinWait from the WinActivate, and uses timeouts, loops and logging to
ensure we have the best chance of success